fix(desktop): require a Codex adapter with Astra support
